Get rid of javax.activation.
Java activation framework (JAF) or javax.activation:activation(maven) is an API that is integral part of
every Java 5 level virtual machine. As such there is no point in still referencing/downloading/including or making any other operation
with it. Simply removing references to it in build scripts makes it use the JVM provided one.

Signed-off-by: Jesse McConnell <jesse.mcconnell@gmail.com>