370387 - SafariWebsocketDraft0Test failure during build.
The reason for the failure of this test was that a BufferedReader was used to read the header lines.
However, the buffered reader may have read and buffered also the hixie bytes and subsequently,
when the test was trying to read the hixie bytes directly from the input stream (and not from the
buffered reader), the read was timing out.
Fixed by always using the input stream to read the header and hixie bytes.