Jetty9 - Moved wiring of SslConnection out of HttpServerConnectionFactory.
The wiring of SslConnection in HttpServerConnectionFactory was not the
right place because a ConnectionFactory should just create the Connection
instances, not also wiring them up.
It is responsibility of the connector to wire the SslConnection in.
Moving the wiring outside HttpServerConnectionFactory fixed also a few
SPDY tests, where the wiring was already done outside (and therefore
was done twice).