jetty-9 - Reorganized SPDY modules.
This change was due to the fact that the old spdy-core and spdy-jetty modules had classes in the same package,
which is a problem for OSGi.
The old spdy-jetty module has been split in 2: spdy-client and spdy-server, and other SPDY modules just underwent a
renaming to make them more sound.