Introduced setQueue(), removed maxQueued() and renamed maxIdleTimeMs() to idleTimeout().
The reason for this change is that CometD needs a way to explicitly set the queue implementation
in load testing to measure latencies.
Property maxQueued was used to decide, at start time, what queue implementation to use, but
now that can be set explicitly with setQueue().
The rename was done because in all other places the idle timeout has been renamed so.