402090 - httpsender PendingState cause uncertain data send to server.
WriteFlusher was storing consumed buffers that may have been reused,
and when the write was being completed those consumed buffer may have
contained new content that was being written too.

Fixed by compacting the buffers at the moment of creation of the
WriteFlusher.PendingState (and not at the moment of the completeWrite() like
it was in historically done, see 4e9460161944b5ccfa4bf1b5494d96ef395a0e76).