409556 Resource files not closed
Made Resource a closeable so that it is easier to close in a try(){} block. Deprecated release() and made it close() instead.

FileResource no longer extends URLResource as it can more efficiently implement all the methods with pure File operations and have no
connections or streams that need to be release/closed