HttpClient optimizations.
HttpConnectionPool now puts "hot" idle connection in the front of a deque,
rather than at the end of a queue, so that the hotter they are the more they
are used.

HttpRequest is now initialized with better defaults so that
HttpConnection.normalizeRequest() has less work to do.
This reduced the number of branches to be executed for each request.

HttpDestination now produces less garbage by using Jetty's BlockingArrayQueue
instead of Java's LinkedBlockingQueue.

HttpClient has now a bigger response buffer, 16 KiB, by default.