445823 Moved RequestLog calling to HttpChannel
Simplified the RequestLogHandler.  It no longer set's async listeners or tries to catch exeptions etc.
Instead it simply sets the RequestLog instance on the HttpChannel.   The HttpChannel#onCompleted method
then uses the log (if set) as the last thing before recycling the connection.  It uses the actual
bytes written (eg after compression) and the actual status sent (eg if sent from a bad message).