Removed incorrect LICENSE file (CREATIVE COMMONS)
This license file was committed to a private repository at github.com:intalio/jetty-fcgi
when we were considering if FCGI was to be a part of Jetty or not. When FCGI was merged into
jetty master (e00f5c40567a91e696e8a5daf881ef595c1adf0d) the file was not removed.

As per the NOTICE.txt file, Jetty is dual licensed under
http://www.apache.org/licenses/LICENSE-2.0.html
and http://www.eclipse.org/legal/epl-v10.html

It was not and is not the intention of the copyright holders or contributors to use the
creative commons license for any part of the jetty release.